+++ src/commands/files/index.ts
@@ -1,3 +1,4 @@
 export { getFile } from './get.js';
+export { deleteFile } from './delete.js';
 export { listFiles, listFilesFormatted, ListFilesOptions } from './list.js';
 export { uploadFile, UploadFileOptions } from './upload.js';